Add has given birth check for changing a Ewe to Ram, disallow it
Former-commit-id: aa3bc9020777d788524a7a87ec87835886902bc4

+++ src/AppBundle/Util/GenderChanger.php
@@ -11,6 +11,7 @@ use AppBundle\Entity\Ram;
 use AppBundle\Enumerator\AnimalObjectType;
 use Doctrine\Common\Persistence\ObjectManager;
 use Doctrine\DBAL\Connection;
+use AppBundle\Component\HttpFoundation\JsonResponse;
 
 /**
  * Class GenderChanger
@@ -108,7 +109,20 @@ class GenderChanger
                 $this->conn->exec($updateQuery);
                 break;
             case AnimalObjectType::Ewe:
-                //Remove relationship from current inheritance table
+
+                //Animal was found, do additional checks to see if we allow a gender change
+                if(count($animal->getBirths()) > 0){
+                    $statusCode = 403;
+                    return new JsonResponse(
+                      array(
+                        Constant::RESULT_NAMESPACE => array (
+                          'code' => $statusCode,
+                          "message" => "Changing the gender of an Ewe which has given birth to a Ram is not allowed for ULN: " . $animal->getUln(),
+                        )
+                      ), $statusCode);
+                }
+                
+                 //Remove relationship from current inheritance table
                 $deleteQuery = "DELETE FROM "  .AnimalObjectType::Ewe ." WHERE id = " .$animal->getId();
                 $this->conn->exec($deleteQuery);
